# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: IncidentIQ - Group Incident Reporting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups/{groupSlug}/reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports for a group (responders/admins with dashboard access)
      parameters:
        - $ref: "#/components/parameters/groupSlug"
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, escalated, resolved]
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, emergency]
        - name: incidentTypeId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated report list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: No dashboard access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createReport
      tags: [reports]
      summary: Submit a new incident report
      parameters:
        - $ref: "#/components/parameters/groupSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportRequest"
      responses:
        "201":
          description: Report created with reference number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupSlug}/reports/{referenceNumber}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a single report with its full activity timeline
      parameters:
        - $ref: "#/components/parameters/groupSlug"
        - name: referenceNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full report with timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupSlug}/reports/{referenceNumber}/updates:
    post:
      operationId: addReportUpdate
      tags: [reports]
      summary: Add a status update, note, or action to a report (responders/admins)
      parameters:
        - $ref: "#/components/parameters/groupSlug"
        - name: referenceNumber
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddReportUpdateRequest"
      responses:
        "201":
          description: Update added to timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportUpdate"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    groupSlug:
      name: groupSlug
      in: path
      required: true
      schema:
        type: string
      description: The unique slug identifying the group
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [error]
...
    Report:
      type: object
      properties:
        id:
          type: string
        referenceNumber:
          type: string
        groupSlug:
          type: string
        incidentTypeId:
          type: string
        incidentTypeName:
          type: string
        severity:
          type: string
          enum: [low, medium, high, emergency]
        description:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        status:
          type: string
          enum: [open, in_progress, escalated, resolved]
        isAnonymous:
          type: boolean
        claimedByUserId:
          type: string
          nullable: true
        claimedByName:
          type: string
          nullable: true
        submittedAt:
          type: string
          format: date-time
        photoCount:
          type: integer
      required: [id, referenceNumber, groupSlug, incidentTypeId, incidentTypeName, severity, description, status, isAnonymous, submittedAt, photoCount]
    ReportDetail:
      type: object
      properties:
        report:
          $ref: "#/components/schemas/Report"
        photos:
          type: array
          items:
            $ref: "#/components/schemas/ReportPhoto"
        updates:
          type: array
          items:
            $ref: "#/components/schemas/ReportUpdate"
        reporter:
          $ref: "#/components/schemas/ReporterInfo"
      required: [report, photos, updates]
    ReporterInfo:
      type: object
      properties:
        name:
          type: string
        userId:
          type: string
          nullable: true
      required: [name]
    ReportPhoto:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        exifTakenAt:
          type: string
          format: date-time
          nullable: true
        exifLatitude:
          type: number
          nullable: true
        exifLongitude:
          type: number
          nullable: true
        uploadedAt:
          type: string
          format: date-time
      required: [id, url, uploadedAt]
    ReportUpdate:
      type: object
      properties:
        id:
          type: string
        updateType:
          type: string
          enum: [claim, status_change, note, photo_added, escalation, auto_claim]
        actorName:
          type: string
        note:
          type: string
          nullable: true
        newStatus:
          type: string
          nullable: true
        escalatedTo:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, updateType, actorName, createdAt]
    CreateReportRequest:
      type: object
      properties:
        incidentTypeId:
          type: string
        severity:
          type: string
          enum: [low, medium, high, emergency]
        description:
          type: string
          minLength: 10
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        isAnonymous:
          type: boolean
          default: false
        photoUrls:
          type: array
          items:
            type: string
          default: []
      required: [incidentTypeId, severity, description]
    AddReportUpdateRequest:
      type: object
      properties:
        updateType:
          type: string
          enum: [claim, status_change, note, escalation]
        note:
          type: string
          nullable: true
        newStatus:
          type: string
          enum: [open, in_progress, escalated, resolved]
          nullable: true
        escalationContactId:
          type: string
          nullable: true
      required: [updateType]
    ReportListResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/Report"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [reports, total, page, limit, totalPages]

    # ── Push Notifications ────────────────────────────────────────────────────
...